try age searching by Stage
git-svn-id: file:///repo/applications/FRED/trunk@11933 976ac4c7-77c1-46cc-9d05-16a3bd20634c

+++ src/nz/cri/gns/fred/util/SampleUtil.java
@@ -571,7 +571,7 @@ public class SampleUtil extends ModelUtil implements FREDConstants, AuditedUtil 
 		return createSample(feature, new Integer(folderId), reuseFeatureAudit, user);
 	}
 	
-	public List<Sample> getSamplesByAge(Integer startAge, Integer stopAge) throws StorageAccessException {
+	public List<Sample> getSamplesByAge(Double startAge, Double stopAge) throws StorageAccessException {
 		Set<Sample> samples = new HashSet<Sample>();
 		List<Stage> stages = fredDAO.getList("FROM Stage AS s WHERE s.baseAge >= ? AND s.topAge <= ?", Stage.class, stopAge, startAge);
 		for (Stage stage : stages) {
